dayPeriod -> slot に変更し整理

+++ server/src/database/enrollments.ts
@@ -1,26 +1,9 @@
 import { PrismaClient } from "@prisma/client";
 import { CourseID, UserID } from "../common/types";
-import { getCoursesWithDayPeriodsByUser } from "./courses";
+import { getCoursesByUserId } from "./courses";
 
 const prisma = new PrismaClient();
 
-// 履修の作成
-export async function createEnrollment({
-  userId,
-  courseId,
-}: {
-  userId: UserID;
-  courseId: CourseID;
-}) {
-  await prisma.enrollment.create({
-    data: {
-      userId,
-      courseId,
-    },
-  });
-}
-
-// 履修の削除
 export async function deleteEnrollment({
   userId,
   courseId,
@@ -36,58 +19,58 @@ export async function deleteEnrollment({
       },
     },
   });
-  const updatedCourses = await getCoursesWithDayPeriodsByUser(userId);
+  const updatedCourses = await getCoursesByUserId(userId);
   return updatedCourses;
 }
 
-// 履修情報の更新
-export async function updateEnrollments({
-  userId,
-  courseId,
-}: {
-  userId: UserID;
-  courseId: CourseID;
-}) {
-  const newCourseDayPeriods = await prisma.courseDayPeriod.findMany({
+/**
+ * `courseId` で与えられた講義の履修を追加する。 `courseId` と同じ曜限の講義が存在する場合はその履修を削除する。
+ */
+export async function createEnrollment(courseId: CourseID, userId: UserID) {
+  // 与えられた講義の履修がすでに存在するかどうかを確認する
+  const targetEnrollment = await prisma.enrollment.findUnique({
     where: {
-      courseId,
-    },
-  });
-  await prisma.$transaction(async (tx) => {
-    // delete current course if exists
-    await tx.enrollment.deleteMany({
-      where: {
+      userId_courseId: {
         userId,
         courseId,
       },
-    });
-    // delete courses in the same day and period as the new course
-    await tx.enrollment.deleteMany({
+    },
+  });
+  // 与えられた講義を履修していない場合のみ、履修の追加を行う
+  if (!targetEnrollment) {
+    const targetCourseSlots = await prisma.slot.findMany({
       where: {
-        userId,
-        course: {
-          courseDayPeriods: {
-            some: {
-              OR: newCourseDayPeriods.map(({ day, period }) => ({
-                AND: [{ day: day }, { period: period }],
-              })),
+        courseId,
+      },
+    });
+    await prisma.$transaction(async (tx) => {
+      // 与えられた講義と同じ曜限の講義の履修を削除する
+      await tx.enrollment.deleteMany({
+        where: {
+          userId,
+          course: {
+            slots: {
+              some: {
+                OR: targetCourseSlots.map(({ day, period }) => ({
+                  AND: [{ day: day }, { period: period }],
+                })),
+              },
             },
           },
         },
-      },
-    });
-    if (courseId !== "") {
-      // add current course
-      await tx.enrollment.create({
-        data: {
-          userId,
-          courseId,
-        },
       });
-    }
-  });
+      if (courseId !== "") {
+        // 与えられた講義の履修を追加する
+        await tx.enrollment.create({
+          data: {
+            userId,
+            courseId,
+          },
+        });
+      }
+    });
+  }
 
-  const updatedCourses = await getCoursesWithDayPeriodsByUser(userId);
-
+  const updatedCourses = await getCoursesByUserId(userId);
   return updatedCourses;
 }
